First look at separating main-thread chunk gen impl from World
this is intended to open the gates to having generators that run on the main thread (e.g. flat), or that use other methods of generating chunks entirely (such as requesting them from a chunk server).
